INTERLOK-3333 Create abstract base class for JSON path service
By having an abstract base class, the common methods can still be shared
between implementations, yet these implementations can set their own
source data input parameters - string vs stream.
