+++ interlok-json/src/main/java/com/adaptris/core/services/path/json/JsonPathService.java
@@ -1,29 +1,17 @@
 package com.adaptris.core.services.path.json;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.EnumSet;
-import java.util.List;
-import java.util.Map;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-import org.apache.commons.lang3.BooleanUtils;
 import com.adaptris.annotation.AdapterComponent;
 import com.adaptris.annotation.AutoPopulated;
 import com.adaptris.annotation.ComponentProfile;
 import com.adaptris.annotation.DisplayOrder;
-import com.adaptris.annotation.InputFieldDefault;
 import com.adaptris.core.AdaptrisMessage;
 import com.adaptris.core.CoreException;
 import com.adaptris.core.ServiceException;
-import com.adaptris.core.ServiceImp;
 import com.adaptris.core.common.Execution;
 import com.adaptris.core.common.StringPayloadDataInputParameter;
-import com.adaptris.core.json.JsonPathExecution;
 import com.adaptris.core.util.ExceptionHelper;
 import com.adaptris.interlok.config.DataInputParameter;
 import com.adaptris.interlok.config.DataOutputParameter;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.Option;
@@ -32,11 +20,16 @@ import com.jayway.jsonpath.ReadContext;
 import com.jayway.jsonpath.spi.json.JsonSmartJsonProvider;
 import com.jayway.jsonpath.spi.mapper.JacksonMappingProvider;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
-import com.thoughtworks.xstream.annotations.XStreamImplicit;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.Setter;
+
+import javax.validation.constraints.NotNull;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.EnumSet;
+import java.util.List;
 
 /**
  * This service allows you to search JSON content and the results are then set back into the message.
@@ -146,7 +139,7 @@ import lombok.Setter;
 @ComponentProfile(summary = "Extract a value from a JSON document", tag = "service,transform,json,metadata")
 @DisplayOrder(order = {"source", "executions", "unwrapJson"})
 @NoArgsConstructor
-public class JsonPathService extends ServiceImp {
+public class JsonPathService extends JsonPathServiceImpl {
 
   /**
    * The source for executing the jsonpath against.
@@ -157,34 +150,9 @@ public class JsonPathService extends ServiceImp {
   @Getter
   @Setter
   @NonNull
-  protected DataInputParameter source = new StringPayloadDataInputParameter();
+  private DataInputParameter<String> source = new StringPayloadDataInputParameter();
 
-  /**
-   * The list of jsonpath executions to apply.
-   *
-   */
-  @NotNull
-  @Valid
-  @AutoPopulated
-  @XStreamImplicit
-  @Getter
-  @Setter
-  @NonNull
-  protected List<Execution> executions = new ArrayList<>();
-
-  protected transient Configuration jsonConfig;
-
-  /**
-   * Get whether the JSON should be unwrapped removing any leading or trailing square brackets
-   * {@code []}.
-   * <p>
-   * The default is false if not specified.
-   * </p>
-   */
-  @Getter
-  @Setter
-  @InputFieldDefault(value = "false")
-  protected Boolean unwrapJson;
+  private transient Configuration jsonConfig;
 
   public JsonPathService(DataInputParameter<String> source, Execution... executions) {
     this(source, new ArrayList<>(Arrays.asList(executions)));
@@ -203,8 +171,7 @@ public class JsonPathService extends ServiceImp {
   public void doService(final AdaptrisMessage message) throws ServiceException {
     try {
 
-      final DataInputParameter<String> src = source;
-      final String rawJson = src.extract(message);
+      final String rawJson = source.extract(message);
       ReadContext context = JsonPath.parse(rawJson, jsonConfig);
 
       for (final Execution execution : executions) {
@@ -231,67 +198,9 @@ public class JsonPathService extends ServiceImp {
     }
   }
 
-  protected static String toString(Object json, Execution exec) throws Exception {
-    ObjectMapper mapper = new ObjectMapper();
-    Object jsonObject = convertIfNull(json, exec);
-    // A JSON Object is effectively a map, so we need to write that out as JSON.
-    // If it's a JSONArray, then "toString" works fine.
-    if (Map.class.isAssignableFrom(jsonObject.getClass())) {
-      return mapper.writeValueAsString(jsonObject);
-    }
-    if (List.class.isAssignableFrom(jsonObject.getClass())) {
-      return mapper.writeValueAsString(jsonObject);
-    }
-    return jsonObject.toString();
-  }
-
-  /**
-   * Strip (if necessary) the leading/trailing [] from the JSON.
-   *
-   * @param json
-   *        The JSON string.
-   */
-  protected static String unwrap(final String json, boolean unwrapJson) {
-    if (unwrapJson) {
-      if (json.startsWith("[") && json.endsWith("]")) {
-        return json.substring(1, json.length() - 1);
-      }
-    }
-    return json;
-  }
-
   @Override
   public void prepare() throws CoreException {
     jsonConfig = new Configuration.ConfigurationBuilder().jsonProvider(new JsonSmartJsonProvider())
         .mappingProvider(new JacksonMappingProvider()).options(EnumSet.noneOf(Option.class)).build();
   }
-
-  @Override
-  protected void closeService() {
-    /* unused/empty method */
-  }
-
-  @Override
-  protected void initService() throws CoreException {
-    // nothing to do.
-  }
-
-  protected boolean unwrapJson() {
-    return BooleanUtils.toBooleanDefaultIfNull(getUnwrapJson(), false);
-  }
-
-  protected boolean suppressPathNotFound(Execution exec) {
-    if (exec instanceof JsonPathExecution) {
-      return ((JsonPathExecution) exec).suppressPathNotFound();
-    }
-    return false;
-  }
-
-  protected static Object convertIfNull(Object o, Execution exec) {
-    if (exec instanceof JsonPathExecution) {
-      return ((JsonPathExecution) exec).nullConverter().convert(o);
-    }
-    return o;
-  }
-
 }
